+++ docs/src/componentDocs/allComponents.mdx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {ComponentCatalogGrids} from '../pages/ComponentCatalogGrids';
+import {ComponentCatalogGrids} from '../components/ComponentCatalogGrids';
 
 {/* KEYWORDS: components typography icon devdoc catalogue catalog */}
