--- docs/src/shared/components/InfoCard/index.ts
@@ -1,1 +1,0 @@
-export * from './InfoCard';
