+++ docs/src/shared/index.ts
@@ -1,3 +1,2 @@
 export * from './utilities';
 export * from './constants';
-export * from './ComponentPreviewTabs';
